defect-RAG/requirements.txt
# Core dependencies
chromadb>=0.5.0
openai>=1.35.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-chroma>=0.1.0

# Web UI
streamlit>=1.38.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
pydantic>=2.0.0
pyyaml>=6.0

# Language detection
langdetect>=1.0.9

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
